move support config from mgradm
