+++ shared/utils/support.go
@@ -5,7 +5,10 @@
 package utils
 
 import (
+	"errors"
 	"fmt"
+	"os/exec"
+	"path"
 	"regexp"
 	"strings"
 	"time"
@@ -31,4 +34,54 @@ func GetSupportConfigFileSaveName() string {
 	}
 	now := time.Now()
 	return fmt.Sprintf("scc_%s_%s", hostname, now.Format("20060102_1504"))
+}
+
+/* CreateSupportConfigTarball will create a tarball in outputFolder with all the supportconfig
+ * files generated by host and pods.
+ */
+func CreateSupportConfigTarball(outputFolder string, files []string) error {
+	// Pack it all into a tarball
+	log.Info().Msg(L("Preparing the tarball"))
+
+	supportFileName := GetSupportConfigFileSaveName()
+	supportFilePath := path.Join(outputFolder, fmt.Sprintf("%s.tar.gz", supportFileName))
+
+	tarball, err := NewTarGz(supportFilePath)
+	if err != nil {
+		return err
+	}
+
+	for _, file := range files {
+		if err := tarball.AddFile(file, path.Join(supportFileName, path.Base(file))); err != nil {
+			return Errorf(err, L("failed to add %s to tarball"), path.Base(file))
+		}
+	}
+	tarball.Close()
+	return nil
+}
+
+// RunSupportConfigOnHost will run supportconfig command on host machine.
+func RunSupportConfigOnHost() ([]string, error) {
+	var files []string
+	extensions := []string{"", ".md5"}
+	// Run supportconfig on the host if installed
+	if _, err := exec.LookPath("supportconfig"); err == nil {
+		out, err := RunCmdOutput(zerolog.DebugLevel, "supportconfig")
+		if err != nil {
+			return []string{}, Errorf(err, L("failed to run supportconfig on the host"))
+		}
+		tarballPath := GetSupportConfigPath(string(out))
+
+		// Look for the generated supportconfig file
+		if tarballPath != "" && FileExists(tarballPath) {
+			for _, ext := range extensions {
+				files = append(files, tarballPath+ext)
+			}
+		} else {
+			return []string{}, errors.New(L("failed to find host supportconfig tarball from command output"))
+		}
+	} else {
+		log.Warn().Msg(L("supportconfig is not available on the host, skipping it"))
+	}
+	return files, nil
 }
